Adding recommendation model 2
This model has been trained for 20 epochs. The results are 90% training and validation accuracy with 0.3046 training loss, 0.3731 validation loss, 0.0826 training mse, and 0.0877 validation mse. 
This model has two inputs, user_id and place_id. From datasets, we pick ratings from user to labeling either the places is recommended or not. If the ratings is < 3, we drop the data and ratings >= 3 will be labeled with value 1 (users preference). Next, we generate negative label (0) randomly from the places that the users never seen. 
Labels are the target we use for train the model. Thus, this model output is a boolean recommendation. We now can use this model to make a recommendation for users with additional code to sorting by ratings or categories that user might be interested in.